+++ libs/py/settings/__init__.py
@@ -1,3 +1,17 @@
-import os
+from pydantic_settings import BaseSettings
+from pydantic import Field, computed_field
 
-LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO").upper()
+
+class LogSettings(BaseSettings):
+    log_level: str = "INFO"
+
+
+class BazelSettings(BaseSettings):
+    workspace: str | None = Field(
+        validation_alias="BUILD_WORKSPACE_DIRECTORY",
+        default=None,
+    )
+
+
+log_settings = LogSettings()
+bazel_settings = BazelSettings()